requirements.txt
selenium==4.15.0
pytest==7.4.3
python-dotenv==1.0.0
pytest-html==4.1.1
webdriver-manager==4.0.1

allure-pytest